LayoutTests/system-preview/rel-ar-setting-enabled-expected.txt


PASS a.relList.supports("ar") should be true when SystemPreviewEnabled is true.
